requirements.txt
# Legacy-compatible requirements for Python 3.10 / chumpy / MANO

numpy==1.23.5
torch
loguru==0.7.3
tyro==0.9.35
loop-rate-limiters==1.2.0
hydra-core==1.3.2
matplotlib==3.10.7
imageio[ffmpeg]==2.37.2
imageio-ffmpeg==0.6.0
opencv-python==4.11.0.86
ipdb==0.13.13
pymeshlab
open3d
chumpy
mujoco-warp @ git+https://github.com/google-deepmind/mujoco_warp@5d5a6450a7699e695f1767efd7715251713da2ac
newton[examples] @ git+https://github.com/newton-physics/newton.git@e15698684b6bfa4c501d49fe938a2837e112380c
